Moved ConditionHelper to sole usage
