+++ FileHelpers/Core/RecordOperations.cs
@@ -3,6 +3,7 @@ using System.Collections;
 using System.Data;
 using System.Reflection;
 using System.Text;
+using FileHelpers.Core;
 using FileHelpers.Helpers;
 
 namespace FileHelpers
